Edit profile description modal
Still need to add the color picker and fix some things. The modal screen does pop up and the tabs does appear but need to fix how it looks because it's a little wonky at the moment.

But, the modal does pop up when the profile picture is clicked which is the intended behaviour.